+++ CHANGELOG.md
@@ -12,6 +12,14 @@
   security updates in June of 2013; continuing to support it would prevent
   the use of newer features of Ruby.
 
+# (unreleased)
+
+* Added support for `ZADD` options `NX`, `XX`, `CH`, `INCR`. See #547.
+
+* Added support for sentinel commands. See #556.
+
+* New `:id` option allows you to identify the client against Redis. See #510.
+
 # 3.2.1
 
 * Added support for `PUBSUB` command.